pyproject.toml
[project]
name = "support-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pytest>=9.0.3",
    "uvicorn>=0.44.0",
    "wandb>=0.25.1",
    "weave>=0.52.36",
]
